chore: update delete workspace section.
