+++ web/ce/components/workspace/index.ts
@@ -1,4 +1,4 @@
 export * from "./edition-badge";
 export * from "./upgrade-badge";
 export * from "./billing";
-export * from "./delete-workspace";
+export * from "./delete-workspace-section";
