+++ web/ee/components/workspace/delete-workspace-section.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/workspace/delete-workspace-section";
