--- web/ee/components/workspace/delete-workspace.tsx
@@ -1,1 +1,0 @@
-export * from "ce/components/workspace/delete-workspace";
